rollout runs without error

+++ day12/data_utils/reader.py
@@ -24,8 +24,7 @@ def get_presents(file_name="input.txt") -> torch.Tensor:
 
         binary_matrix: list[list[int]] = [[1 if char == '#' else 0 for char in row]
                                           for row in split_matrix]
-        # Convert to PyTorch tensor with float16 dtype
-        tensor_matrix = torch.tensor(binary_matrix, dtype=torch.uint8)
+        tensor_matrix = torch.tensor(binary_matrix, dtype=torch.float32)
         extracted_present_tensors.append(tensor_matrix)
 
     return torch.stack(extracted_present_tensors)
@@ -57,5 +56,5 @@ def get_placement_info(file_name="input.txt") -> list[PlacementInfo]:
         present_count_str = present_count_str.split()
         present_count = list(map(int, present_count_str))
         args.append(PlacementInfo(width, height, torch.tensor(
-            present_count, dtype=torch.int64)))
+            present_count, dtype=torch.float32)))
     return args
